fix e2e shell bug: can not find ./hack/run-os-compitable-e2e.sh rename compability to compatibility
