+++ test/kubean_os_compatibility_e2e/e2e-install-cluster/hosts-conf-cm.yml
