+++ test/kubean_os_compatibility_e2e/e2e-install-cluster/kubeanClusterOps.yml
