+++ test/kubean_os_compatibility_e2e/kubean_suite_test.go
@@ -1,4 +1,4 @@
-package kubean_oscompability_e2e
+package kubean_os_compatibility_e2e
 
 import (
 	"testing"
